enterprise-rag-eval/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0.0
tqdm>=4.66.0
pydantic>=2.0.0
pytest>=7.0.0
openai>=1.0.0
